Refactored replacevar data collector
