+++ js/src/replacevar/data-collector.js
@@ -1,0 +1,10 @@
+/**
+ * Prototype class for replacevar data collectors.
+ */
+class DataCollector {
+	getParentTitle() {
+		return "";
+	}
+}
+
+export default DataCollector;
