+++ js/src/replacevar/index.js
@@ -1,0 +1,4 @@
+export { default as DataProvider } from "./data-provider";
+export { default as GutenbergDataCollector } from "./gutenberg-data-collector";
+export { default as TinyMceDataCollector } from "./tiny-mce-data-collector";
+export { default as DataCollector } from "./data-collector";
